corrected mistakes

+++ src/js/03-feedback.js
@@ -14,16 +14,24 @@ refs.form.addEventListener('submit', onSubmit);
 const userData = {};
 
 function onTextInput(evt) {
+    console.log(evt.target.value)
     userData[evt.target.name] = evt.target.value;
     localStorage.setItem(LOCALSTORAGE_KEY, JSON.stringify(userData));
 };
 
 function onSubmit(evt) {
-    console.log(userData);
-
     evt.preventDefault();
-    evt.target.reset();
-    localStorage.clear();
+
+    const {elements : {email, message}} = evt.target;
+    if(email.value === '' || message.value === '') {
+        alert('Fill out ALL the fields');
+    }
+
+    else {
+        console.log(userData);
+        evt.target.reset();
+        localStorage.clear();
+    }
 }
 
 function onPageReload() {
@@ -32,6 +40,10 @@ function onPageReload() {
     if(savedInfo) {
         refs.email.value = savedInfo.email;
         refs.message.value = savedInfo.message;
+    }
+    else {
+        refs.email.value = '';
+        refs.message.value = '';
     }
 }
